requirements.txt
fastapi
uvicorn
pytest
python-dotenv
supabase
langchain
langchain-openai
httpx
